chore: update setting
