+++ frontend/src/components/GeneralSetting/index.ts
@@ -2,9 +2,8 @@ import AIAugmentationSetting from "./AIAugmentationSetting.vue";
 import AccountSetting from "./AccountSetting.vue";
 import AnnouncementSetting from "./AnnouncementSetting.vue";
 import BrandingSetting from "./BrandingSetting.vue";
-import DatabaseChangeModeSetting from "./DatabaseChangeModeSetting.vue";
+import GeneralSetting from "./GeneralSetting.vue";
 import MaximumSQLResultSizeSetting from "./MaximumSQLResultSizeSetting.vue";
-import NetworkSetting from "./NetworkSetting.vue";
 import ProductImprovementSetting from "./ProductImprovementSetting.vue";
 import SecuritySetting from "./SecuritySetting.vue";
 
@@ -13,9 +12,8 @@ export {
   SecuritySetting,
   AccountSetting,
   MaximumSQLResultSizeSetting,
-  NetworkSetting,
   AIAugmentationSetting,
   AnnouncementSetting,
-  DatabaseChangeModeSetting,
+  GeneralSetting,
   ProductImprovementSetting,
 };
